fix wrong cursor position when insert empty style
If there is a '\n' in a style's {after} property, the cursor will be at
a wrong line (below the right one) after insertion.
For example, inserting a codeblock will leave the cursor at the last
line with '```'.
After this fix, the cursor will be between the two lines of '```',
namely, in the middle.